requirements.txt
# Flask Web 框架
Flask==3.0.0

# HTTP 请求
requests==2.31.0

# 环境变量管理
python-dotenv==1.0.0

# 跨域支持（生产环境可选）
Flask-CORS==4.0.0

# API 限流（生产环境可选）
Flask-Limiter==3.5.0

# Gunicorn WSGI 服务器（生产环境）
gunicorn==21.2.0

# 等待依赖（可选，用于健康检查）
waitress==3.0.0
